chore: Update package dependencies
This commit updates the package dependencies in `package.json` and `yarn.lock`.

In `package.json`, the following dependency has been added:
- `@swc/core`: "^1.3.71"

In `yarn.lock`, various versions of `@swc/core` and its optional dependencies have been resolved.

These updates are essential for keeping the project up-to-date and ensuring compatibility with the latest versions of the dependencies.